Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpReceiver-getParameters-expected.txt


FAIL getParameters() with audio receiver assert_equals: Expect dictionary.rtcp to be plain object expected "object" but got "undefined"
FAIL getParameters() with video receiver assert_equals: Expect dictionary.rtcp to be plain object expected "object" but got "undefined"
FAIL getParameters() with simulcast video receiver assert_equals: Expect dictionary.rtcp to be plain object expected "object" but got "undefined"
